requirements.txt
openai==1.109.1
python-dotenv==1.0.1
streamlit==1.51.0

#library for images with streamlit
pillow==11.3.0
